requirements.txt
Django>=4.0,<=4.1
djangorestframework>=3.13,<=3.14
pytest
pytest-cov
pytest-djangoapp